requirements.txt
requests
pandas
numpy
scipy
python-dotenv
psycopg2-binary
schedule
nflreadpy

